move base test class to test's __init__
